Support tilde expansion ("~") in file paths.
Without this patch, org-krita supports absolute and relative
filepaths, but the unzip process doesn't successfully support the
tilde in a test Debian Buster system.

Fortunately, we can expand the tilde before giving it to unzipper to
prevent any filepath issues.

Now, the following three links all display in the org-mode buffer and
are all equivalent:

- [[krita:krita-test.kra][krita-test]]
- [[krita:/home/test/krita-test.kra][krita-test]]
- [[krita:~/krita-test.kra][krita-test]]